+++ app/controllers/admin_controller.rb
@@ -1,0 +1,6 @@
+class AdminController < ApplicationController
+  before_action :authenticate_user!
+  before_action :check_system_admin
+
+
+end