+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Enhancements
 - **Administrative Changes**
-  - Fixed a minor typo
+  - Added an administrative dashboard and a link in the dropdown menu for system admins
 - **Dataset Changes**
   - Dataset file access is now controlled by the new DAUA process